% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_crispr.R
\name{sample_noise}
\alias{sample_noise}
\title{Generate Constant Random Noise Across sgRNAs}
\usage{
sample_noise(SD, n_total)
}
\arguments{
\item{SD}{Standard deviation of the noise.}

\item{n_total}{Total number of sgRNAs to apply the noise to.}
}
\value{
A numeric vector of length `n_total` with identical noise values.
}
\description{
This function generates a constant random noise value drawn from a normal distribution with mean 0 and standard deviation `SD`,
and repeats it across all sgRNAs. It is used to simulate global or shared perturbations across all guides.
}
