% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_seq.R
\name{seq_add}
\alias{seq_add}
\title{Simulate Amplification and Sequencing Process}
\usage{
seq_add(
  wide_data,
  rounds = 5,
  totalDepth = 1e+09,
  cm_mu = 0.99,
  cm_sd = 0.01,
  ep_mu = 0.01,
  ep_sd = 0.001,
  pcr_sd = 0.02,
  sf_sd = 0.03
)
}
\arguments{
\item{wide_data}{A data frame of sgRNA counts before amplification, typically from a prior simulation step.}

\item{rounds}{Number of PCR amplification rounds to simulate. Default is 5.}

\item{totalDepth}{Total sequencing depth to distribute across sgRNAs. Default is \code{1e9}.}

\item{cm_mu}{Mean of the capture efficiency \eqn{\mu_{\text{cm}}} used in generating \eqn{\mu_{\text{PCR}}}.}

\item{cm_sd}{Standard deviation \eqn{\sigma_{\text{cm}}} of the capture efficiency.}

\item{ep_mu}{Mean of the error propagation rate during PCR.}

\item{ep_sd}{Standard deviation of the error propagation rate.}

\item{pcr_sd}{Standard deviation \eqn{\sigma_{\text{PCR}}} used in the diagonal covariance matrix \eqn{\Sigma_{\text{PCR}}}.}

\item{sf_sd}{Standard deviation for sampling fluctuation during the sequencing process.}
}
\value{
A data frame of simulated read counts after sequencing, incorporating amplification and sequencing biases.
}
\description{
This function simulates the sequencing output by applying amplification and sequencing steps to sgRNA count data.
It models biases and variability introduced by PCR amplification, capture efficiency, and sequencing depth.
}
